+++ CHANGELOG.md
@@ -27,6 +27,12 @@
 
 -->
 
+## Unreleased
+
+### Added
+
+- `ReadParam` for `isize` and `usize`
+
 ## v0.0.1 - 2026/04/10
 
 First release.